make HTTP parser usable by external alien forces
